requirements.txt
flask==2.3.3
firebase-admin==6.2.0
requests==2.26.0
pyrebase4==4.5.0
requests==2.31.0
jinja2==3.1.2
Flask-Mail==0.9.1